Mark the CEnum trait as #[doc(hidden)]
After spending a little bit thinking about this trait, it turns out that
it is a pretty big backwards compatibility risk for any crate that
happens to the c-enum crate. Any major update to the c-enum crate now
requires a breaking change from any crate that uses it since downstream
crates can interact with the CEnum trait.