Make variants without explicit values properly follow C enum rules
In C, enum variants without an explicit value take the value $prev + 1,
or 0, if they are the first variant. The constants generated by this
crate are meant to follow the same convention.

Before this commit, due to a bug in the macro, they don't follow that
convention. What happens instead is that all variants without assigned
values have the value 0, unless they directly follow a variant with an
assigned value, in which case they take the same value as the previous
variant. This is _wrong_. The fix here is to explicitly pass a default
value through and set that to either 0 (for the first field) or
Self::$prev + 1 (for all other fields).

I have also included test cases that verify that the values of enum
variants are as expected.