Match multiple fields at a time in order to support larger enums
Since the c_enum! implementation is recursive an enum with lots of
variants can quickly result in hitting rustc's recursion limit. We
can avoid this by reducing recursion.

This commit does so by generating a bunch of intermediate macro variants
which consume 8, 32, or 128 fields at a time. That way, an enum with a
large number of variants requires much fewer recursive macro invocations
because its fields are handled in batches and only the last few fields
are handled individually.

Writing these extra matchers out by hand is rather painful so I have
extracted them out to a separate file and written a python script to
generate it.